+++ src/Pages/Login/LoginForm/components/index.ts
@@ -1,0 +1,4 @@
+export { LoginInput } from './Input';
+export { LoginPanel } from './Panel';
+export { UserIdentifyForm } from './UserIdentifyForm';
+export { LoginLogo } from './Logo';
